extract building hours stateful list from the index file
